Tests/LibWeb/Text/expected/css/keyframes-find-rule.txt

last matching rule: true
matching selector list: 25%, 75%
matching reordered selector list: 75%, 25%
to matches 100%: 100%
partial selector list: null
invalid selectors return null: true
